+++ sql/_00_instantiate-schema.sql
@@ -99,14 +99,34 @@ CREATE TABLE media (
     rss_source rss_source,
     uploader VARCHAR(25),
     -- metadata pertaining to the media item
+    -- - other identifiers
     imdb_id VARCHAR(10) CHECK (imdb_id ~ '^tt[0-9]{7,8}$'),
     tmdb_id INTEGER CHECK (tmdb_id > 0),
+    -- - quantitative details
+    budget BIGINT CHECK (budget >= 0),
+    revenue BIGINT CHECK (revenue >= 0),
+    runtime INTEGER CHECK (runtime >= 0),
+    -- - country and production information
+    origin_country CHAR(2)[],
+    production_companies VARCHAR(255)[],
+    production_countries CHAR(2)[],
+    production_status VARCHAR(25),
+    -- - language information
+    original_language CHAR(2),
+    spoken_languages CHAR(2)[],
+    -- - other string fields
     genre VARCHAR(20)[],
-    language VARCHAR(20)[],
+    original_media_title VARCHAR(255),
+    -- - long string fields
+    tagline VARCHAR(255),
+    overview TEXT,
+    -- - ratings info
+    tmdb_rating DECIMAL(5,3) CHECK (tmdb_rating BETWEEN 0 AND 10),
+    tmdb_votes INTEGER CHECK (tmdb_votes >= 0),
     rt_score INTEGER CHECK (rt_score IS NULL OR (rt_score BETWEEN 0 AND 100)),
     metascore INTEGER CHECK (metascore IS NULL OR (metascore BETWEEN 0 AND 100)),
-    imdb_rating DECIMAL(3,1) CHECK (imdb_rating IS NULL OR (imdb_rating BETWEEN 0 AND 100)),
-    imdb_votes INTEGER,
+    imdb_rating DECIMAL(4,1) CHECK (imdb_rating IS NULL OR (imdb_rating BETWEEN 0 AND 100)),
+    imdb_votes INTEGER CHECK (imdb_votes >= 0),
     -- metadata pertaining to the video file
     resolution VARCHAR(10),
     video_codec VARCHAR(10),
@@ -235,14 +255,34 @@ COMMENT ON COLUMN media.original_link IS 'may contain either the direct download
 COMMENT ON COLUMN media.rss_source IS 'source of rss feed for item ingestion, if any';
 COMMENT ON COLUMN media.uploader IS 'uploading entity of the media item';
 -- metadata pertaining to the media item
-COMMENT ON COLUMN media.imdb_id IS 'IMDB identifier for media item';
-COMMENT ON COLUMN media.tmdb_id IS 'identifier for themoviedb.org API';
-COMMENT ON COLUMN media.genre IS 'array of genres associated with the movie';
-COMMENT ON COLUMN media.language IS 'array of languages available encoded in ISO 639 format';
-COMMENT ON COLUMN media.rt_score IS 'Rotten Tomatoes score';
-COMMENT ON COLUMN media.metascore IS 'MetaCritic score';
-COMMENT ON COLUMN media.imdb_rating IS 'IMDB rating out of 10';
-COMMENT ON COLUMN media.imdb_votes IS 'number of votes on IMDB';
+-- - other id fields
+COMMENT ON COLUMN media.imdb_id IS 'from TMDB; IMDB identifier for media item';
+COMMENT ON COLUMN media.tmdb_id IS 'from TMDB; identifier for themoviedb.org API';
+-- - quantitative details
+COMMENT ON COLUMN media.budget IS 'from TMDB; production budget of media item';
+COMMENT ON COLUMN media.revenue IS 'from TMDB; current revenue of media item';
+COMMENT ON COLUMN media.runtime IS 'from TMDB; runtime in minutes of media item';
+-- - country and production information
+COMMENT ON COLUMN media.origin_country IS 'from TMDB; primary country of production in iso_3166_1 format';
+COMMENT ON COLUMN media.production_companies IS 'from TMDB; array of production companies';
+COMMENT ON COLUMN media.production_countries IS 'from TMDB; array of countries where media item was produced in iso_3166_1 format';
+COMMENT ON COLUMN media.production_status IS 'from TMDB; current production status of media item';
+-- - language information
+COMMENT ON COLUMN media.original_language IS 'from TMDB; primary language of media item in ISO 639 format';
+COMMENT ON COLUMN media.spoken_languages IS 'from TMDB; array of languages available encoded in ISO 639 format';
+-- - other string fields
+COMMENT ON COLUMN media.genre IS 'from TMDB; array of genres associated with the movie';
+COMMENT ON COLUMN media.original_media_title IS 'from TMDB; original title of media item';
+-- - long string fields
+COMMENT ON COLUMN media.tagline IS 'from TMDB; tagline for the media item';
+COMMENT ON COLUMN media.overview IS 'from TMDB; brief plot synopsis of media item';
+-- - ratings info
+COMMENT ON COLUMN media.tmdb_rating IS 'from TMDB; rating submitted by TMDB users out of 10';
+COMMENT ON COLUMN media.tmdb_votes IS 'from TMDB; number of ratings by TMDB users';
+COMMENT ON COLUMN media.rt_score IS 'from OMDb; Rotten Tomatoes score out of 100';
+COMMENT ON COLUMN media.metascore IS 'from OMDb; MetaCritic score out of 100';
+COMMENT ON COLUMN media.imdb_rating IS 'from OMDb; IMDB rating out of 100';
+COMMENT ON COLUMN media.imdb_votes IS 'from OMDb; number of votes on IMDB';
 -- metadata pertaining to the video file
 COMMENT ON COLUMN media.resolution IS 'video resolution';
 COMMENT ON COLUMN media.video_codec IS 'video compression codec';